Updated By Avdhut With ResponsEntity In project

+++ EnquiryDetails/src/main/java/com/finance/repository/EnquiryRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.finance.model.Enquiry;
 
 @Repository
-public interface EnquiryRepository extends JpaRepository<Enquiry, Integer>
+public interface EnquiryRepository extends JpaRepository<Enquiry, Long>
 	{
-
+		
 	}
